Add share card entry points to mobile forecast details
